+++ src/lib/mrvVk/mrvTimelineViewport.cpp
@@ -77,7 +77,22 @@ namespace mrv
             }
         }
 
+        const timeline::ShaderOptions& TimelineViewport::getShaderOptions() const noexcept
+        {
+            return _p->shaderOptions;
+        }
 
+        void TimelineViewport::setShaderOptions(const timeline::ShaderOptions& value) noexcept
+        {
+            TLRENDER_P();
+            
+            if (value == p.shaderOptions)
+                return;
+            
+            p.shaderOptions = value;
+            redrawWindows();
+        }
+        
         void TimelineViewport::_showPixelBar() const noexcept
         {
             TLRENDER_P();
@@ -176,8 +191,6 @@ namespace mrv
                 0.F, static_cast<float>(viewportSize.h), -1.F, 1.F);
             return pm * vm;
         }
-        
-